[update:NA] update.
Description:
update.

Major Changes:
1. change VALUE_TYPE_STRING to VALUE_TYPE_STRING_POINTER.
2. code format.

+++ tests/core/test_array_stack.c
@@ -80,20 +80,20 @@ static int test_array_stack_pop(TEST_ENTRY *entry)
 
         expect_value = 7;
         stack->pop(stack, (void **)&p);
-        THROW_IF(p != expect_value, -1);
+        THROW_IF((int)p != expect_value, -1);
 
         expect_value = 6;
         stack->pop(stack, (void **)&p);
-        THROW_IF(p != expect_value, -1);
+        THROW_IF((int)p != expect_value, -1);
         dbg_str(DBG_VIP, "%x:%x", (int)p, expect_value);
 
         expect_value = 5;
         stack->pop(stack, (void **)&p);
-        THROW_IF(p != expect_value, -1);
+        THROW_IF((int)p != expect_value, -1);
 
         expect_value = 4;
         stack->pop(stack, (void **)&p);
-        THROW_IF(p != expect_value, -1);
+        THROW_IF((int)p != expect_value, -1);
     } CATCH (ret) {} FINALLY {
         object_destroy(stack);
     }
